Move ensureSingleEditor and calcIsReversed along with targetsToContinuousTarget
