+++ packages/cursorless-engine/src/processTargets/TargetPipelineRunner.ts
@@ -20,6 +20,8 @@ import { ContainingTokenIfUntypedEmptyStage } from "./modifiers/ConditionalModif
 import { PlainTarget } from "./targets";
 import { uniqWithHash } from "../util/uniqWithHash";
 import {targetsToContinuousTarget} from "../util/targetsToContinuousTarget";
+import {ensureSingleEditor} from "../util/targetsToContinuousTarget";
+import {calcIsReversed} from "../util/targetsToContinuousTarget";
 
 export class TargetPipelineRunner {
   constructor(
@@ -278,28 +280,12 @@ function getExcludedScope(
   );
 }
 
-export function calcIsReversed(anchor: Target, active: Target) {
-  if (anchor.contentRange.start.isAfter(active.contentRange.start)) {
-    return true;
-  }
-  if (anchor.contentRange.start.isBefore(active.contentRange.start)) {
-    return false;
-  }
-  return anchor.contentRange.end.isAfter(active.contentRange.end);
-}
-
 function uniqTargets(array: Target[]): Target[] {
   return uniqWithHash(
     array,
     (a, b) => a.isEqual(b),
     (a) => a.contentRange.concise(),
   );
-}
-
-export function ensureSingleEditor(anchorTarget: Target, activeTarget: Target) {
-  if (anchorTarget.editor !== activeTarget.editor) {
-    throw new Error("Cannot form range between targets in different editors");
-  }
 }
 
 function targetsToVerticalTarget(
